Moved AccessToken from azure-core to azure-identity. Polished Javadoc for identity classes.
