+++ sdk/identity/azure-identity/src/main/java/com/azure/android/identity/AccessToken.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.credential;
+package com.azure.android.identity;
 
 import org.threeten.bp.OffsetDateTime;
