+++ sdk/identity/azure-identity/src/main/java/com/azure/android/identity/TokenRequestHandle.java
@@ -1,7 +1,5 @@
 package com.azure.android.identity;
 
-import com.azure.android.core.credential.AccessToken;
-
 import org.threeten.bp.Duration;
 
 import java.util.ArrayList;
@@ -10,33 +8,33 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 /**
- * A Handle that a background running thread (e.g. OkHttp thread) uses to synchronously receive an access token
- * produced by the main UI thread. A background running thread can initiate the access token request by calling
- * {@link TokenRequestObservable#sendRequest(List)} ()}, which returns a {@link TokenRequestHandle} instance. The
- * background thread then waits for the arrival of the access token by calling
- * {@link TokenRequestHandle#waitForToken(Duration)}.
+ * A {@link TokenRequestHandle handle} that a background running thread (e.g. OkHttp thread) uses to synchronously
+ * receive an {@link AccessToken} produced by the main UI thread. A background running thread can initiate the
+ * {@link AccessToken} request by calling {@link TokenRequestObservable#sendRequest(List)} ()}, which returns a
+ * {@link TokenRequestHandle} instance. The background thread then waits for the arrival of the {@link AccessToken}
+ * by calling {@link TokenRequestHandle#waitForToken(Duration)}.
  * <p>
  * IMPLEMENTATION NOTE: The {@link TokenRequestObservable} INTERNALLY uses the same {@link TokenRequestHandle} object
  * (returned from {@link TokenRequestObservable#sendRequest(List)}) as an Event to notify a {@link TokenRequestObserver}
- * about the need for access token. Upon receiving this event, the {@link TokenRequestObserver} invokes
- * {@link TokenRequestObserver#onTokenRequest(String[], TokenResponseCallback)} (TokenResponseCallback)} which is
- * responsible for retrieving the access token, the observer then sets the access token or access token retrieval
- * error in the Handle using it's package PRIVATE setters {@link TokenRequestHandle#setToken(AccessToken)},
- * {@link TokenRequestHandle#setError(Throwable)}.
+ * about the need for {@link AccessToken}. Upon receiving this event, the {@link TokenRequestObserver} invokes
+ * {@link TokenRequestObserver#onTokenRequest(String[], TokenResponseCallback)} ({@link TokenResponseCallback})}
+ * which is responsible for retrieving the {@link AccessToken}, the observer then sets the {@link AccessToken} or
+ * {@link AccessToken} retrieval error in the Handle using it's package PRIVATE setters
+ * {@link TokenRequestHandle#setToken(AccessToken)}, {@link TokenRequestHandle#setError(Throwable)}.
  * <p>
  * It is important that an Event instance is processed only once by the {@link TokenRequestObserver}. The
- * {@link TokenRequestObservable} uses {@link androidx.lifecycle.LiveData} to post the Event to
- * {@link TokenRequestObserver}, the LiveData is designed to cache the latest value (in this case Event) and
- * re-deliver when there is a new observer. An example for such unexpected re-delivery is: Let's say the
- * {@link TokenRequestObservable} is owned by a {@link androidx.lifecycle.ViewModel}, the ViewModel lives through
- * multiple Activity::onCreate - Activity::onDestroy spans https://developer.android
- * .com/topic/libraries/architecture/viewmodel#lifecycle. Let's say in Activity::onCreate(Bundle) we register a
- * {@link TokenRequestObserver} to this ViewModel scoped Observable. When a background thread request for token, the
- * observer receives the request and process it. If user begins rotating the screen (i.e configuration changed) then
- * the Activity::onDestroy() is called and once the rotation is done the Activity::onCreate(Bundle) method will be
- * called again, which result in registration of new {@link TokenRequestObserver} and internally LiveData re-deliver
- * the old cached/handled event, which we definitely don't want the new Observer to handle. To identify such an
- * Event, the Observer uses {@link TokenRequestHandle#isConsumed()}.
+ * {@link TokenRequestObservable} uses a {@link androidx.lifecycle.LiveData} instance to post the Event to
+ * {@link TokenRequestObserver}. {@link androidx.lifecycle.LiveData} is designed to cache the latest value (in
+ * this case Event) and re-deliver it when there is a new Observer. An example for unexpected re-delivery is: Let's
+ * say the {@link TokenRequestObservable} is owned by a {@link androidx.lifecycle.ViewModel} which lives
+ * through multiple {@code Activity::onCreate} - {@code Activity::onDestroy} spans and in
+ * {@code Activity::onCreate} we register a {@link TokenRequestObserver} to this {@link androidx.lifecycle.ViewModel}
+ * 's scoped Observable. When a background thread requests for a token, the Observer will receive the request and
+ * process it. If the user begins rotating the screen (i.e configuration changed) then {@code Activity::onDestroy} is
+ * called and once the rotation is done {@code Activity::onCreate} method will be called again, which will result in
+ * the registration of a new {@link TokenRequestObserver} and will internally cause
+ * {@link androidx.lifecycle.LiveData} to re-deliver the old cached/handled Event, which we definitely don't want the
+ * new Observer to handle. To identify such an Event, the Observer uses {@link TokenRequestHandle#isConsumed()}.
  */
 class TokenRequestHandle {
     // The scope of the token requested by the handle owner.
@@ -52,7 +50,7 @@ class TokenRequestHandle {
     private Throwable error;
 
     /**
-     * Creates TokenRequestHandle.
+     * Creates a {@link TokenRequestHandle}.
      *
      * @param scopes The requested token scope.
      */
@@ -61,15 +59,16 @@ class TokenRequestHandle {
     }
 
     /**
-     * Synchronously wait for the arrival of an access token requested via
+     * Synchronously wait for the arrival of an {@link AccessToken} requested via
      * {@link TokenRequestObservable#sendRequest(List)}}.
      *
      * @param timeout The maximum time to wait in milliseconds.
-     * @return The access token.
-     * @throws Throwable
+     * @return The {@link AccessToken}.
+     * @throws Throwable if there is an error.
      */
     public AccessToken waitForToken(Duration timeout) throws Throwable {
         this.latch.await(timeout.toMillis(), TimeUnit.MILLISECONDS);
+
         if (this.error != null) {
             throw this.error;
         } else {
@@ -104,10 +103,10 @@ class TokenRequestHandle {
     /**
      * PACKAGE PRIVATE METHOD.
      *
-     * Sets the access token retrieved and unblocks the background thread waiting in
+     * Sets the {@link AccessToken} retrieved and unblocks the background thread waiting in
      * {@link TokenRequestHandle#waitForToken(Duration)}.
      *
-     * @param token The access token.
+     * @param token The {@link AccessToken}.
      */
     public void setToken(AccessToken token) {
         if (this.token != null) {
@@ -121,7 +120,7 @@ class TokenRequestHandle {
     /**
      * PACKAGE PRIVATE METHOD.
      *
-     * Sets the error during access token retrieval and unblocks the background thread waiting in
+     * Sets the error during {@link AccessToken} retrieval and unblocks the background thread waiting in
      * {@link TokenRequestHandle#waitForToken(Duration)}.
      *
      * @param error The error.
