+++ sdk/identity/azure-identity/src/main/java/com/azure/android/identity/TokenResponseCallback.java
@@ -1,20 +1,18 @@
 package com.azure.android.identity;
 
-import com.azure.android.core.credential.AccessToken;
-
 /**
  * The callback that {@link TokenRequestObserver} uses to notify the result of token retrieval.
  */
 public interface TokenResponseCallback {
     /**
-     * Notify the successful retrieval of an access token.
+     * Notify the successful retrieval of an {@link AccessToken}.
      *
      * @param token The retrieved token.
      */
     void onToken(AccessToken token);
 
     /**
-     * Notify the error occurred during access token retrieval.
+     * Notify the error occurred during {@link AccessToken} retrieval.
      *
      * @param t The error.
      */
